+++ redis/src/client.rs
@@ -3,14 +3,18 @@ use std::time::Duration;
 #[cfg(feature = "aio")]
 use crate::aio::AsyncPushSender;
 use crate::{
-    connection::{connect, Connection, ConnectionInfo, ConnectionLike, IntoConnectionInfo},
+    connection::{
+        connect, connect_with_io, io, Connection, ConnectionInfo, ConnectionLike,
+        IntoConnectionInfo,
+    },
     types::{RedisResult, Value},
+    GenericConnection,
 };
 #[cfg(feature = "aio")]
 use std::pin::Pin;
 
 #[cfg(feature = "tls-rustls")]
-use crate::tls::{inner_build_with_tls, TlsCertificates};
+use crate::connection::io::tls_rustls::{inner_build_with_tls, TlsCertificates};
 
 #[cfg(feature = "cache-aio")]
 use crate::caching::CacheConfig;
@@ -56,6 +60,14 @@ impl Client {
         connect(&self.connection_info, None)
     }
 
+    /// Same as [`Self::get_connection`] but with a custom IO Driver.
+    pub fn get_connection_with_io<IO: io::ConnectionDriver>(
+        &self,
+        driver: IO,
+    ) -> RedisResult<GenericConnection<IO>> {
+        connect_with_io(driver, &self.connection_info, None)
+    }
+
     /// Instructs the client to actually connect to redis with specified
     /// timeout and returns a connection object.  The connection object
     /// can be used to send commands to the server.  This can fail with
@@ -63,6 +75,15 @@ impl Client {
     /// that you handle those errors.
     pub fn get_connection_with_timeout(&self, timeout: Duration) -> RedisResult<Connection> {
         connect(&self.connection_info, Some(timeout))
+    }
+
+    /// Same as [`Self::get_connection_with_timeout`] but with a custom IO Driver.
+    pub fn get_connection_with_io_and_timeout<IO: io::ConnectionDriver>(
+        &self,
+        driver: IO,
+        timeout: Duration,
+    ) -> RedisResult<GenericConnection<IO>> {
+        connect_with_io(driver, &self.connection_info, Some(timeout))
     }
 
     /// Returns a reference of client connection info object.
