+++ redis/src/connection/info.rs
@@ -1,0 +1,267 @@
+use std::{borrow::Cow, str::FromStr};
+
+use crate::types::HashMap;
+use crate::{ErrorKind, ProtocolVersion, RedisError, RedisResult};
+
+use super::ConnectionAddr;
+
+/// This function takes a redis URL string and parses it into a URL
+/// as used by rust-url.
+///
+/// This is necessary as the default parser does not understand how redis URLs function.
+pub fn parse_redis_url(input: &str) -> Option<url::Url> {
+    match url::Url::parse(input) {
+        Ok(result) => match result.scheme() {
+            "redis" | "rediss" | "redis+unix" | "unix" => Some(result),
+            _ => None,
+        },
+        Err(_) => None,
+    }
+}
+
+/// Holds the connection information that redis should use for connecting.
+#[derive(Clone, Debug)]
+pub struct ConnectionInfo {
+    /// A connection address for where to connect to.
+    pub addr: ConnectionAddr,
+
+    /// A redis connection info for how to handshake with redis.
+    pub redis: RedisConnectionInfo,
+}
+
+/// Redis specific/connection independent information used to establish a connection to redis.
+#[derive(Clone, Debug, Default)]
+pub struct RedisConnectionInfo {
+    /// The database number to use.  This is usually `0`.
+    pub db: i64,
+    /// Optionally a username that should be used for connection.
+    pub username: Option<String>,
+    /// Optionally a password that should be used for connection.
+    pub password: Option<String>,
+    /// Version of the protocol to use.
+    pub protocol: crate::ProtocolVersion,
+}
+
+impl FromStr for ConnectionInfo {
+    type Err = RedisError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        s.into_connection_info()
+    }
+}
+
+/// Converts an object into a connection info struct.  This allows the
+/// constructor of the client to accept connection information in a
+/// range of different formats.
+pub trait IntoConnectionInfo {
+    /// Converts the object into a connection info object.
+    fn into_connection_info(self) -> RedisResult<ConnectionInfo>;
+}
+
+impl IntoConnectionInfo for ConnectionInfo {
+    fn into_connection_info(self) -> RedisResult<ConnectionInfo> {
+        Ok(self)
+    }
+}
+
+/// URL format: `{redis|rediss}://[<username>][:<password>@]<hostname>[:port][/<db>]`
+///
+/// - Basic: `redis://127.0.0.1:6379`
+/// - Username & Password: `redis://user:password@127.0.0.1:6379`
+/// - Password only: `redis://:password@127.0.0.1:6379`
+/// - Specifying DB: `redis://127.0.0.1:6379/0`
+/// - Enabling TLS: `rediss://127.0.0.1:6379`
+/// - Enabling Insecure TLS: `rediss://127.0.0.1:6379/#insecure`
+/// - Enabling RESP3: `redis://127.0.0.1:6379/?protocol=resp3`
+impl IntoConnectionInfo for &str {
+    fn into_connection_info(self) -> RedisResult<ConnectionInfo> {
+        match parse_redis_url(self) {
+            Some(u) => u.into_connection_info(),
+            None => fail!((ErrorKind::InvalidClientConfig, "Redis URL did not parse")),
+        }
+    }
+}
+
+impl<T> IntoConnectionInfo for (T, u16)
+where
+    T: Into<String>,
+{
+    fn into_connection_info(self) -> RedisResult<ConnectionInfo> {
+        Ok(ConnectionInfo {
+            addr: ConnectionAddr::Tcp(self.0.into(), self.1),
+            redis: RedisConnectionInfo::default(),
+        })
+    }
+}
+
+/// URL format: `{redis|rediss}://[<username>][:<password>@]<hostname>[:port][/<db>]`
+///
+/// - Basic: `redis://127.0.0.1:6379`
+/// - Username & Password: `redis://user:password@127.0.0.1:6379`
+/// - Password only: `redis://:password@127.0.0.1:6379`
+/// - Specifying DB: `redis://127.0.0.1:6379/0`
+/// - Enabling TLS: `rediss://127.0.0.1:6379`
+/// - Enabling Insecure TLS: `rediss://127.0.0.1:6379/#insecure`
+/// - Enabling RESP3: `redis://127.0.0.1:6379/?protocol=resp3`
+impl IntoConnectionInfo for String {
+    fn into_connection_info(self) -> RedisResult<ConnectionInfo> {
+        match parse_redis_url(&self) {
+            Some(u) => u.into_connection_info(),
+            None => fail!((ErrorKind::InvalidClientConfig, "Redis URL did not parse")),
+        }
+    }
+}
+
+fn parse_protocol(query: &HashMap<Cow<str>, Cow<str>>) -> RedisResult<ProtocolVersion> {
+    Ok(match query.get("protocol") {
+        Some(protocol) => {
+            if protocol == "2" || protocol == "resp2" {
+                ProtocolVersion::RESP2
+            } else if protocol == "3" || protocol == "resp3" {
+                ProtocolVersion::RESP3
+            } else {
+                fail!((
+                    ErrorKind::InvalidClientConfig,
+                    "Invalid protocol version",
+                    protocol.to_string()
+                ))
+            }
+        }
+        None => ProtocolVersion::RESP2,
+    })
+}
+
+const DEFAULT_PORT: u16 = 6379;
+
+pub(super) fn url_to_tcp_connection_info(url: url::Url) -> RedisResult<ConnectionInfo> {
+    let host = match url.host() {
+        Some(host) => {
+            // Here we manually match host's enum arms and call their to_string().
+            // Because url.host().to_string() will add `[` and `]` for ipv6:
+            // https://docs.rs/url/latest/src/url/host.rs.html#170
+            // And these brackets will break host.parse::<Ipv6Addr>() when
+            // `client.open()` - `ActualConnection::new()` - `addr.to_socket_addrs()`:
+            // https://doc.rust-lang.org/src/std/net/addr.rs.html#963
+            // https://doc.rust-lang.org/src/std/net/parser.rs.html#158
+            // IpAddr string with brackets can ONLY parse to SocketAddrV6:
+            // https://doc.rust-lang.org/src/std/net/parser.rs.html#255
+            // But if we call Ipv6Addr.to_string directly, it follows rfc5952 without brackets:
+            // https://doc.rust-lang.org/src/std/net/ip.rs.html#1755
+            match host {
+                url::Host::Domain(path) => path.to_string(),
+                url::Host::Ipv4(v4) => v4.to_string(),
+                url::Host::Ipv6(v6) => v6.to_string(),
+            }
+        }
+        None => fail!((ErrorKind::InvalidClientConfig, "Missing hostname")),
+    };
+    let port = url.port().unwrap_or(DEFAULT_PORT);
+    let addr = if url.scheme() == "rediss" {
+        #[cfg(any(feature = "tls-native-tls", feature = "tls-rustls"))]
+        {
+            match url.fragment() {
+                Some("insecure") => ConnectionAddr::TcpTls {
+                    host,
+                    port,
+                    insecure: true,
+                    tls_params: None,
+                },
+                Some(_) => fail!((
+                    ErrorKind::InvalidClientConfig,
+                    "only #insecure is supported as URL fragment"
+                )),
+                _ => ConnectionAddr::TcpTls {
+                    host,
+                    port,
+                    insecure: false,
+                    tls_params: None,
+                },
+            }
+        }
+
+        #[cfg(not(any(feature = "tls-native-tls", feature = "tls-rustls")))]
+        fail!((
+            ErrorKind::InvalidClientConfig,
+            "can't connect with TLS, the feature is not enabled"
+        ));
+    } else {
+        ConnectionAddr::Tcp(host, port)
+    };
+    let query: HashMap<_, _> = url.query_pairs().collect();
+    Ok(ConnectionInfo {
+        addr,
+        redis: RedisConnectionInfo {
+            db: match url.path().trim_matches('/') {
+                "" => 0,
+                path => path.parse::<i64>().map_err(|_| -> RedisError {
+                    (ErrorKind::InvalidClientConfig, "Invalid database number").into()
+                })?,
+            },
+            username: if url.username().is_empty() {
+                None
+            } else {
+                match percent_encoding::percent_decode(url.username().as_bytes()).decode_utf8() {
+                    Ok(decoded) => Some(decoded.into_owned()),
+                    Err(_) => fail!((
+                        ErrorKind::InvalidClientConfig,
+                        "Username is not valid UTF-8 string"
+                    )),
+                }
+            },
+            password: match url.password() {
+                Some(pw) => match percent_encoding::percent_decode(pw.as_bytes()).decode_utf8() {
+                    Ok(decoded) => Some(decoded.into_owned()),
+                    Err(_) => fail!((
+                        ErrorKind::InvalidClientConfig,
+                        "Password is not valid UTF-8 string"
+                    )),
+                },
+                None => None,
+            },
+            protocol: parse_protocol(&query)?,
+        },
+    })
+}
+
+#[cfg(unix)]
+pub(super) fn url_to_unix_connection_info(url: url::Url) -> RedisResult<ConnectionInfo> {
+    let query: HashMap<_, _> = url.query_pairs().collect();
+    Ok(ConnectionInfo {
+        addr: ConnectionAddr::Unix(url.to_file_path().map_err(|_| -> RedisError {
+            (ErrorKind::InvalidClientConfig, "Missing path").into()
+        })?),
+        redis: RedisConnectionInfo {
+            db: match query.get("db") {
+                Some(db) => db.parse::<i64>().map_err(|_| -> RedisError {
+                    (ErrorKind::InvalidClientConfig, "Invalid database number").into()
+                })?,
+
+                None => 0,
+            },
+            username: query.get("user").map(|username| username.to_string()),
+            password: query.get("pass").map(|password| password.to_string()),
+            protocol: parse_protocol(&query)?,
+        },
+    })
+}
+
+#[cfg(not(unix))]
+pub(super) fn url_to_unix_connection_info(_: url::Url) -> RedisResult<ConnectionInfo> {
+    fail!((
+        ErrorKind::InvalidClientConfig,
+        "Unix sockets are not available on this platform."
+    ));
+}
+
+impl IntoConnectionInfo for url::Url {
+    fn into_connection_info(self) -> RedisResult<ConnectionInfo> {
+        match self.scheme() {
+            "redis" | "rediss" => url_to_tcp_connection_info(self),
+            "unix" | "redis+unix" => url_to_unix_connection_info(self),
+            _ => fail!((
+                ErrorKind::InvalidClientConfig,
+                "URL provided is not a redis URL"
+            )),
+        }
+    }
+}
