+++ redis/src/sentinel.rs
@@ -117,15 +117,17 @@ use crate::aio::MultiplexedConnection as AsyncConnection;
 use crate::client::AsyncConnectionConfig;
 
 use crate::{
-    connection::ConnectionInfo, types::RedisResult, Client, Cmd, Connection, ErrorKind,
-    FromRedisValue, IntoConnectionInfo, RedisConnectionInfo, TlsMode, Value,
+    connection::{ActualConnection, ConnectionInfo},
+    types::RedisResult,
+    Client, Cmd, Connection, ConnectionDriver, ErrorKind, FromRedisValue, GenericConnection,
+    IntoConnectionInfo, RedisConnectionInfo, TlsMode, Value,
 };
 
 /// The Sentinel type, serves as a special purpose client which builds other clients on
 /// demand.
-pub struct Sentinel {
+pub struct Sentinel<IO = ActualConnection> {
     sentinels_connection_info: Vec<ConnectionInfo>,
-    connections_cache: Vec<Option<Connection>>,
+    connections_cache: Vec<Option<GenericConnection<IO>>>,
     #[cfg(feature = "aio")]
     async_connections_cache: Vec<Option<AsyncConnection>>,
     replica_start_index: usize,
@@ -399,20 +401,21 @@ async fn async_try_single_sentinel<T: FromRedisValue>(
     }
 }
 
-fn reconnect(
-    connection: &mut Option<Connection>,
+fn reconnect<IO: ConnectionDriver>(
+    connection: &mut Option<GenericConnection<IO>>,
     connection_info: &ConnectionInfo,
 ) -> RedisResult<()> {
     let sentinel_client = Client::open(connection_info.clone())?;
-    let new_connection = sentinel_client.get_connection()?;
+    let io = IO::establish_conn(&connection_info.addr, None)?;
+    let new_connection = sentinel_client.get_connection_with_io(io)?;
     connection.replace(new_connection);
     Ok(())
 }
 
-fn try_single_sentinel<T: FromRedisValue>(
+fn try_single_sentinel<T: FromRedisValue, IO: ConnectionDriver>(
     cmd: Cmd,
     connection_info: &ConnectionInfo,
-    cached_connection: &mut Option<Connection>,
+    cached_connection: &mut Option<GenericConnection<IO>>,
 ) -> RedisResult<T> {
     if cached_connection.is_none() {
         reconnect(cached_connection, connection_info)?;
@@ -763,6 +766,15 @@ impl SentinelClient {
     pub fn get_connection(&mut self) -> RedisResult<Connection> {
         let client = self.get_client()?;
         client.get_connection()
+    }
+
+    /// Same as [`Self::get_connection`] but with your own IO driver.
+    pub fn get_connection_with_io<IO: ConnectionDriver>(
+        &mut self,
+        driver: IO,
+    ) -> RedisResult<GenericConnection<IO>> {
+        let client = self.get_client()?;
+        client.get_connection_with_io(driver)
     }
 }
